Delete the browser registry, and move artist cleanup to /tools/artists
ADR-0081 points 3, 4 and 5. Net -596 lines.

**Point 4 first, because it unblocks point 3.** Artist cleanup was a "browser"
at /library/artist-cleanup, routed through the registry. It is a job you run
against the library — what Tools means under ADR-0058 point 2 — so it is now an
ordinary screen at /tools/artists beside Duplicates, Artwork and Organiser.
That removed the registry's only routed member.

**Point 3.** What was left was one registered browser whose only consumer
imported it directly, bypassing the registry entirely. Gone: LibraryBrowser,
LibraryView, browsers/index.ts, browserRegistry, registerBrowser, getBrowser,
getBrowsers, getBrowsersByCategory, BrowserMetadata, RegisteredBrowser,
DEFAULT_BROWSER_ID, and routes.ts's BROWSER_ROUTES, PARKED_BROWSERS and
LIBRARY_ITEMS — the last two of which were already empty once artist cleanup
left.

**Point 5.** DiscoverBrowser is now components/Embed/DiscoverSurface.tsx, beside
the only thing that renders it, and takes the two props it actually reads rather
than the twenty-four a browser took. EmbedDiscover used to supply every field,
most of them empty, so that a field the surface started reading was a type error
at the seam instead of undefined inside a WKWebView. That guard is kept and
improves: reading a third field is now a compile error at the definition.

**components/Library/ is not deleted, and I deleted it once by mistake.** Point
3 names symbols, not a directory, and that directory also holds columnDefinitions,
AlphabetBar, TrackContextMenu, AlbumContextMenu, SelectionIndicator and the
shared filter and context-menu types — all of which components/shared/ and
hooks/ import, and point 6 says shared/ stays because the embed path runs
through it. Restored; only the registry left.

**The navigation test lost the half that never caught anything.** Its
registry-based suite checked three lists against each other; the surviving half
reads App.tsx and checks that every Link resolves to a mounted route. Its own
docstring records why — /favorites and /downloads sat in the sidebar through the
ADR-0057 strip, silently redirecting, because they were hardcoded in a component
rather than declared in a list. It earned that again here: this change left a
Link to /library/artist-cleanup in LibraryPage.tsx pointing at a route that no
longer existed.

Verified: tsc reports the same 6 pre-existing errors as origin/main and no new
ones, 364 unit tests pass, the web build succeeds, and DiscoverSurface is still
its own 84 kB chunk so the embedded page the Apple clients load is unchanged in
shape.
