--- packages/frontend/src/components/Library/browsers/DiscoverBrowser/index.ts
@@ -1,17 +1,0 @@
-import { lazy } from 'react';
-import { registerBrowser } from '../../types';
-
-const LazyDiscoverBrowser = lazy(() => import('./DiscoverBrowser'));
-
-registerBrowser(
-  {
-    id: 'discover',
-    name: 'Discover',
-    description: 'New releases, recommendations, and music to explore',
-    icon: 'Sparkles',
-    category: 'discovery',
-    requiresFeatures: false,
-    requiresEmbeddings: false,
-  },
-  LazyDiscoverBrowser
-);
