--- packages/frontend/src/components/Library/index.ts
@@ -1,5 +1,0 @@
-export { LibraryView } from './LibraryView';
-export { LibraryBrowser } from './LibraryBrowser';
-export { SelectionIndicator } from './SelectionIndicator';
-export { TrackContextMenu } from './TrackContextMenu';
-export * from './types';
